parallelized refactor done
